Move everything to separate class
