+++ lib/run_loop/core_simulator.rb
@@ -46,6 +46,9 @@ class RunLoop::CoreSimulator
   attr_reader :xcrun
 
   # @!visibility private
+  attr_reader :sim_keyboard
+
+  # @!visibility private
   METADATA_PLIST = '.com.apple.mobile_container_manager.metadata.plist'
 
   # @!visibility private
@@ -401,6 +404,11 @@ class RunLoop::CoreSimulator
   end
 
   # @!visibility private
+  def sim_keyboard
+    @sim_keyboard ||= RunLoop::SimKeyboardSettings.new(device)
+  end
+
+  # @!visibility private
   def simctl
     @simctl ||= RunLoop::Simctl.new
   end
@@ -424,7 +432,7 @@ class RunLoop::CoreSimulator
 
     RunLoop::CoreSimulator.quit_simulator
     RunLoop::CoreSimulator.ensure_hardware_keyboard_connected(pbuddy)
-    device.simulator_ensure_software_keyboard_will_show
+    sim_keyboard.simulator_ensure_software_keyboard_will_show
 
     args = ['open', '-g', '-a', sim_app_path, '--args',
             '-CurrentDeviceUDID', device.udid,
@@ -802,7 +810,7 @@ Command had no output.
       return true
     end
 
-    if !device.simulator_software_keyboard_will_show?
+    if !sim_keyboard.simulator_software_keyboard_will_show?
       RunLoop.log_debug("Simulator relaunch required:  software keyboard is minimized")
       return true
     end
