+++ lib/run_loop/device.rb
@@ -40,7 +40,6 @@ module RunLoop
     attr_reader :state
     attr_reader :simulator_root_dir
     attr_reader :simulator_accessibility_plist_path
-    attr_reader :simulator_preferences_plist_path
     attr_reader :simulator_log_file_path
 
     # Create a new device.
@@ -326,14 +325,6 @@ version: #{version}
     end
 
     # @!visibility private
-    def simulator_preferences_plist_path
-      return nil if physical_device?
-
-      directory = File.join(simulator_root_dir, "data", "Library", "Preferences")
-      pbuddy.ensure_plist(directory, "com.apple.Preferences.plist")
-    end
-
-    # @!visibility private
     def simulator_log_file_path
       return nil if physical_device?
       @simulator_log_file_path ||= File.join(CORE_SIMULATOR_LOGS_DIR, udid,
@@ -534,27 +525,6 @@ Could not update the Simulator languages.
       end
 
       running_apps
-    end
-
-    # Check if all the properties needed for the soft keyboard to appear are set
-    # Approach to negate 'true' and 'false' was chosen in order to do not reboot sim too often
-    # as this will cover the cases when the properties are not set which means that keyboard will be shown anyways
-    def simulator_software_keyboard_will_show?
-      plist = simulator_preferences_plist_path
-
-      hw_keyboard_disabled = pbuddy.plist_read('HardwareKeyboardLastSeen', plist) != 'true'
-      minimization_disabled = pbuddy.plist_read('AutomaticMinimizationEnabled', plist) != 'true'
-      soft_keyboard_enabled = pbuddy.plist_read('SoftwareKeyboardShownByTouch', plist) != 'false'
-
-      hw_keyboard_disabled && minimization_disabled && soft_keyboard_enabled
-    end
-
-    # Add properties needed for soft keyboard to show into preferences plist
-    def simulator_ensure_software_keyboard_will_show
-      plist = simulator_preferences_plist_path
-      pbuddy.plist_set('HardwareKeyboardLastSeen', 'bool', 'NO', plist)
-      pbuddy.plist_set('SoftwareKeyboardShownByTouch', 'bool', 'YES', plist)
-      pbuddy.plist_set('AutomaticMinimizationEnabled', 'bool', 'NO', plist)
     end
 
 =begin
