+++ spec/lib/device_spec.rb
@@ -618,47 +618,6 @@ describe RunLoop::Device do
     end
   end
 
-  context "software keyboard will show" do
-    let(:keyboard_enabled) { Resources.shared.plist_with_software_keyboard(true) }
-    let(:keyboard_not_enabled) { Resources.shared.plist_with_software_keyboard(false) }
-    let(:simulator) { RunLoop::Device.new("denis", "9.0", "udid") }
-
-    context "#simulator_software_keyboard_will_show?" do
-      it "returns true if Preferences.plist:AutomaticMinimizationEnabled is 0" do
-        expect(simulator).to(
-          receive(:simulator_preferences_plist_path).and_return(keyboard_enabled)
-        )
-
-        expect(simulator.simulator_software_keyboard_will_show?).to be == true
-      end
-
-      it "returns false if Preferences.plist:AutomaticMinimizationEnabled is not 0" do
-        expect(simulator).to(
-          receive(:simulator_preferences_plist_path).and_return(keyboard_not_enabled)
-        )
-
-        expect(simulator.simulator_software_keyboard_will_show?).to be == false
-      end
-    end
-
-    it "#simulator_ensure_software_keyboard_will_show" do
-      plist = File.join(Resources.shared.local_tmp_dir, "Preferences.plist")
-      FileUtils.rm_rf(plist)
-      FileUtils.cp(keyboard_not_enabled, plist)
-
-      expect(simulator).to(
-        receive(:simulator_preferences_plist_path).at_least(:once).and_return(plist)
-      )
-
-      expect(simulator.simulator_software_keyboard_will_show?).to be == false
-
-      actual = simulator.simulator_ensure_software_keyboard_will_show
-      expect(actual).to be_truthy
-
-      expect(simulator.simulator_software_keyboard_will_show?).to be == true
-    end
-  end
-
   describe 'updating the device state' do
 
     before do
